+++ History.md
@@ -1,8 +1,14 @@
+unreleased
+==================
+* added: pid property and announcePid so calling process can know about child's pid [bo01ean](https://github.com/bo01ean)
+
 1.18.1 / 2015-05-18
+==================
 
 * changed: Added io.js support [rwky](https://github.com/rwky)
 
 1.18.0 / 2015-05-18
+==================
 
 * changed: Removed support for node 0.8 and added support for 0.12 [rwky](https://github.com/rwky)
 * changed: Listen to stdin error event for spawn errors [kapouer](https://github.com/kapouer)